requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
clickhouse-driver==0.2.7
cryptography==42.0.5
Deprecated==1.2.14
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
regex==2024.4.16
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.29
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.2
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.1
